requirements.txt
streamlit==1.32.0
openai==1.6.1
python-dotenv==1.0.1
flask==2.3.2
gunicorn==21.2.0   # optional, needed only for deploying Flask on Render

#pip install -r requirements.txt
#pip freeze > requirements.txt